refactor: RecruitApplicantBar 디렉토리 이동
